Add build instructions
I suppose that these are obvious for someone who is familiar
with Crystal; but I wasn't, and it took me a while to figure
out how to build the code to test my changes :)

I've also added a working_dir to the Compose file, so that
the user is placed directly in the source directory. I hope
this won't affect other uses (e.g. inside VSCode).