Fix SSH when using non-default port + add test harness
Since Ubuntu 22.10, SSH uses socket activation. This means that the
SSH server isn't running by default, and gets started automatically
the first time there is a connection on port 22.

Upside: it saves 3 MB of RAM.

Downside: if you're customizing the SSH port number with a drop-in
configuration file at cloud-init time, it breaks SSH. SSH then needs
to be restarted (or the machine needs to be rebooted) for the port
number to be picked up by the systemd generator. The net result for
hetzner-k3s is that if we use a non-default SSH port, provisioning
breaks.

This manifests itself by the following log message:

  systemd[1]: ssh.socket: Socket unit configuration has changed while
  unit has been running, no open socket file descriptor left. The
  socket unit is not functional until restarted.

One possible fix is to disable socket activation and revert to the
default mode (start SSH at boot). This can be done by disabling
ssh.socket and enabling ssh.service.

This patch does exactly that, adding the corresponding commands to
the cloud-init template. This has been tested with Ubuntu 24.04
and 22.04 as well.

The following link has more details on the Ubuntu change:

https://discourse.ubuntu.com/t/sshd-now-uses-socket-based-activation-ubuntu-22-10-and-later/30189/14

This patch also adds a test harness to automatically create a
bunch of clusters with different configurations and verify that
they work correctly. It was helpful to confirm that this patch
worked correctly with all the distros available on Hetzner; or
rather, that the only ones that didn't work (alma-8, centos-8...)
weren't working in the first place anyway.