fix: cloud init did not properly run, contained bashisms
At least for autoscaled nodes the working install script
is run as part of cloud init, i.e. MUST be POSIX compliant.

We changed it for the master installer as well, since, while it
works for distris with default ssh shell bash it would also fail,
for minimal distris, with only /bin/sh