fix: dns name of api server should not be in kubeconfig
From the docs:

> # api_server_hostname: k8s.example.com # optional: DNS for the k8s API LoadBalancer. After the script has run, create a DNS record with the address of the API LoadBalancer.

The problem is, that the installer will hang forever at the
next command, which is kubectl cluster-info, when the DNS name
is in the kubeconfig.

Since, the user has no chance to configure that DNS name, not knowing
the IP of the created api loadbalancer at installer run time.

He would have to ctrl-c the script at that point, create the DNS entry
by checking hetzner UI for loadbalancer IP and then rerun the script.

=> I find it way better to have the LB *IP* in the kubeconfig, then the
installer can run through.

User can at any time later change it to dns name if wanted, i.e. when he
configured DNS and it'll work, the certification is there (tls-sans is set).